avance en proyecto
avance en proyecto

+++ frontend/pedidos/src/app/components/log-in/log-in.component.ts
@@ -1,3 +1,6 @@
+import { AlertDialogComponent } from './../alert-dialog/alert-dialog.component';
+import { NgxSpinnerService } from 'ngx-spinner';
+import { MatDialog } from '@angular/material/dialog';
 import { AuthService } from './../../servicios/auth.service';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
@@ -19,17 +22,50 @@ export class LogInComponent implements OnInit {
   hide = true;
   constructor(
     private fb: FormBuilder,
-    private authService:AuthService,    private router: Router) { }
+    private authService:AuthService,    private router: Router,    private spinner: NgxSpinnerService,
+    public dialog: MatDialog) { }
 
   ngOnInit(): void {
+    localStorage.removeItem("token");
   }
 
-  onSubmit() {
-      // TODO: Use EventEmitter with form value
-      console.log(this.loginForm.value);
-    }
+
 
     validarUsuario() {
+      this.spinner.show();
+
+
+      if(!this.loginForm.value.username){
+        this.dialog.open(AlertDialogComponent, {
+          data: {
+            title: '',
+            text: `Por favor ingrese el usuario`,
+            icon: 'error',
+            showCancelButton: false,
+            showConfirmButton: true,
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar',
+          },
+        });
+        this.spinner.hide();
+        return;
+      }
+
+      if(!this.loginForm.value.password){
+        this.dialog.open(AlertDialogComponent, {
+          data: {
+            title: '',
+            text: `Por favor ingrese contraseña`,
+            icon: 'error',
+            showCancelButton: false,
+            showConfirmButton: true,
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar',
+          },
+        });
+        this.spinner.hide();
+        return;
+      }
       let password =crypto.SHA512(this.loginForm.value.password).toString();
      const login = {
         usuario: this.loginForm.value.username,
@@ -37,11 +73,24 @@ export class LogInComponent implements OnInit {
       };
 
       this.authService.singin(login).subscribe((res:any)=>{
-        console.log(res);
-        localStorage.setItem('token',res.token);
-        this.router.navigate(['cliente']);
+        if(!res.respuesta){
+     this.dialog.open(AlertDialogComponent, {
+            data: {
+              title: '',
+              text: `Usuario o contraseña invalidos`,
+              icon: 'error',
+              showCancelButton: false,
+              showConfirmButton: true,
+              confirmButtonText: 'Aceptar',
+              cancelButtonText: 'Cancelar',
+            },
+          });
+        }else{
+          localStorage.setItem('token',res.token);
+          this.router.navigate(['cliente']);
+        }
      });
-
+     this.spinner.hide();
     }
